Fixed some warnings related to unsigned-to-signed conversions. (#520) Added prototype for set_csv_options. (#519)
git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@2461 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ tape/histogram.cpp
@@ -172,7 +172,7 @@ int parse_bin_enum(char *cptr, BIN *bin, PROPERTY *prop){
 	for(kw = prop->keywords; kw != NULL; kw = kw->next){
 		if(strcmp(kw->name, pos) == 0){
 			bin->low_inc = bin->high_inc = 1;
-			bin->low_val = bin->high_val = kw->value;
+			bin->low_val = bin->high_val = (double)(kw->value);
 			return 1;
 		}
 	}
